Reorganize constants and data classes
