--- app/src/main/java/data/CalendarEntry.kt
@@ -1,9 +1,0 @@
-package data
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class CalendarEntry(
-    val id: Long = -1,
-    val displayName: String = "Select a calendar..."
-)
