--- app/src/main/java/data/ContractData.kt
@@ -1,10 +1,0 @@
-package data
-
-import ei.Ei.ContractCoopStatusResponse
-import ei.Ei.LocalContract
-
-// Data class used as return object from api.fetchContractData
-data class ContractData(
-    val contracts: List<LocalContract>,
-    val contractStatuses: List<ContractCoopStatusResponse>
-)
