+++ app/src/main/java/data/Missions.kt
@@ -1,6 +1,13 @@
 package data
 
+import ei.Ei.MissionInfo
 import kotlinx.serialization.Serializable
+
+// Data class used as return object from api.fetchMissionData
+data class MissionData(
+    val missions: List<MissionInfo>,
+    val virtueMissions: List<MissionInfo>
+)
 
 // Data class used to save mission information to preferences
 @Serializable
