+++ app/src/main/java/data/Tanks.kt
@@ -1,0 +1,17 @@
+package data
+
+import kotlinx.serialization.Serializable
+
+// Data class used to save fuel tank info to preferences
+@Serializable
+data class TankInfo(
+    var level: Int = 0,
+    var fuelLevels: List<FuelLevelInfo> = emptyList()
+)
+
+@Serializable
+data class FuelLevelInfo(
+    var eggId: Int,
+    var fuelQuantity: Double,
+    var fuelSlider: Double
+)
