+++ app/src/main/java/data/constants/Api.kt
@@ -1,0 +1,10 @@
+package data.constants
+
+const val API_ROOT = "https://www.auxbrain.com"
+const val MISSION_ENDPOINT = "${API_ROOT}/ei_afx/get_active_missions_v2"
+const val BACKUP_ENDPOINT = "${API_ROOT}/ei/bot_first_contact"
+const val CONTRACT_ENDPOINT = "${API_ROOT}/ei/coop_status_bot"
+const val PERIODICALS_ENDPOINT = "${API_ROOT}/ei/get_periodicals"
+const val CONTRACTS_ARCHIVE_ENDPOINT = "${API_ROOT}/ei_ctx/get_contracts_archive"
+
+const val CURRENT_CLIENT_VERSION = 99
