+++ app/src/main/java/data/constants/CustomEggs.kt
@@ -1,0 +1,9 @@
+package data.constants
+
+// Tiers for colleggtible rewards
+val FARM_SIZE_TIERS = arrayOf(
+    10_000_000.0, // Tier 1: 10M chickens
+    100_000_000.0, // Tier 2: 100M chickens
+    1_000_000_000.0, // Tier 3: 1B chickens
+    10_000_000_000.0, // Tier 4: 10B chickens
+)
