+++ app/src/main/java/data/constants/Missions.kt
@@ -1,0 +1,26 @@
+package data.constants
+
+val ALL_SHIPS = arrayOf(
+    "afx_ship_chicken_1",
+    "afx_ship_chicken_9",
+    "afx_ship_chicken_heavy",
+    "afx_ship_bcr",
+    "afx_ship_millenium_chicken",
+    "afx_ship_corellihen_corvette",
+    "afx_ship_galeggtica",
+    "afx_ship_defihent",
+    "afx_ship_voyegger",
+    "afx_ship_henerprise",
+    "afx_ship_atreggies"
+)
+
+val TANK_SIZES = arrayOf(
+    2000000000,
+    200000000000,
+    10000000000000,
+    100000000000000,
+    200000000000000,
+    300000000000000,
+    400000000000000,
+    500000000000000
+)
