+++ app/src/main/java/data/constants/Notifications.kt
@@ -1,0 +1,3 @@
+package data.constants
+
+const val CONTRACT_NOTIFICATION_CHANNEL_ID = "widgetegg_contract_notifications"
