+++ app/src/main/java/data/constants/UI.kt
@@ -1,0 +1,51 @@
+package data.constants
+
+import androidx.compose.ui.graphics.Color
+
+val DEFAULT_WIDGET_BACKGROUND_COLOR = Color(0xff181818)
+val DEFAULT_WIDGET_TEXT_COLOR = Color.White
+
+val NUMBER_UNITS = arrayOf(
+    "k",
+    "M",
+    "B",
+    "T",
+    "q",
+    "Q",
+    "s",
+    "S",
+    "o",
+    "N",
+    "d",
+    "U",
+    "D",
+    "Td",
+    "qd",
+    "Qd",
+    "sd",
+    "Sd",
+    "Od",
+    "Nd",
+    "V",
+    "uV",
+    "dV",
+    "tV",
+    "qV",
+    "QV",
+    "sV",
+    "SV",
+    "OV",
+    "NV",
+    "tT"
+)
+
+val PROBLEMATIC_BROWSERS = arrayOf(
+    "org.mozilla.firefox",
+    "com.duckduckgo.mobile.android"
+)
+
+const val DEFAULT_BROWSER = "com.android.chrome"
+
+const val PROGRESS_BACKGROUND_COLOR = "#464646"
+const val CONTRACT_PROGRESS_COLOR = "#008531"
+const val CONTRACT_OFFLINE_PROGRESS_COLOR = "#51dda8"
