+++ app/src/main/java/widget/contracts/large/ContractWidgetLarge.kt
@@ -41,15 +41,15 @@ import androidx.glance.state.PreferencesGlanceStateDefinition
 import androidx.glance.text.Text
 import androidx.glance.text.TextStyle
 import androidx.glance.unit.ColorProvider
-import data.CONTRACT_OFFLINE_PROGRESS_COLOR
-import data.CONTRACT_PROGRESS_COLOR
+import data.constants.CONTRACT_OFFLINE_PROGRESS_COLOR
+import data.constants.CONTRACT_PROGRESS_COLOR
 import data.ContractInfoEntry
 import data.ContributorInfoEntry
-import data.DEFAULT_BROWSER
-import data.DEFAULT_WIDGET_BACKGROUND_COLOR
-import data.DEFAULT_WIDGET_TEXT_COLOR
-import data.PROBLEMATIC_BROWSERS
-import data.PROGRESS_BACKGROUND_COLOR
+import data.constants.DEFAULT_BROWSER
+import data.constants.DEFAULT_WIDGET_BACKGROUND_COLOR
+import data.constants.DEFAULT_WIDGET_TEXT_COLOR
+import data.constants.PROBLEMATIC_BROWSERS
+import data.constants.PROGRESS_BACKGROUND_COLOR
 import data.PeriodicalsContractInfoEntry
 import data.SeasonGradeAndGoals
 import kotlinx.coroutines.CoroutineScope
